KIALI-2429 WIP - skeleton for mtls endpoint
