+++ handlers/tls.go
@@ -1,0 +1,33 @@
+package handlers
+
+import (
+	"net/http"
+
+	"github.com/gorilla/mux"
+
+	"github.com/kiali/kiali/business"
+	"github.com/kiali/kiali/log"
+)
+
+// NamespaceTls is the API to get namespace-wide mTLS status
+func NamespaceTls(w http.ResponseWriter, r *http.Request) {
+	params := mux.Vars(r)
+
+	// Get business layer
+	business, err := business.Get()
+	if err != nil {
+		RespondWithError(w, http.StatusInternalServerError, "Services initialization error: "+err.Error())
+		return
+	}
+
+	namespace := params["namespace"]
+
+	status, err := business.TLS.NamespaceWidemTLSStatus(namespace)
+	if err != nil {
+		log.Error(err)
+		RespondWithError(w, http.StatusInternalServerError, err.Error())
+		return
+	}
+
+	RespondWithJSON(w, http.StatusOK, status)
+}
